+++ utils.hpp
@@ -35,7 +35,10 @@ int num_digits_array(long long unsigned* ints, int ints_size);
  * @return      TRUE if the concatenation of the ints in the array is
  *              pandigital for the supplied range, FALSE otherwise.
  */
-bool is_pandigital_for_range(int start, int end, size_t* ints, int ints_size);
+bool is_pandigital_for_range(size_t start,
+                             size_t end,
+                             size_t* ints,
+                             size_t ints_size);
 
 
 #endif // UTILS_HPP_
